v1.0 Button polish
Split ButtonTagHelper into two; one that generates button elements and
another that generates links ('a' tags). This is so we can accurately
add OutputElementHint attributes and get better intellisense, which is
very useful for this particular component.

Removed some attributes that can now just be passed through directly
(now that intellisense is improved by the above change) - type, name,
value etc.

Also removed the use of LinkTagHelperBase. Turns out we can extend the
built-in tag helpers that do link generation for our own elements and
remove a tonne of complexity.