Overhaul page template imports to work with CSP
All the previous mechanisms for automatically importing scripts, calling
initialization scripts etc. are incompatible with CSP.

The const members on HtmlSnippets are now deprecated in favour of methods
on a new type - PageTemplateHelper. The methods here allow for passing a
CSP nonce. In addition this type also provides hashes of the two inline
scripts.

The Options type has a new member for users to specify how to get the
CSP nonce for the current request.

The Tag helper component and Razor page template are updated to use
PageTemplateHelper and the new options.