Make date model binding more flexible
An IModelBinderProvider is configured that will attempt to model bind
all Date, DateTime and DateOnly instances in a way that works with the
Date input component. However, sometimes apps pass dates around without
using the Date input component (e.g. as a query parameter) and these
model binders are blocking the built-in ones from running and working.

This exposes a method on our options class to create one of our model
binders. With this, consumers mix in our model binders with their own as
they see fit.