Revise types used for attributes and collections
In the component Options classes, switch to IReadOnlyCollection<T>
(instead of T[]) and IReadOnlyDictionary<string, string?> (instead of
Dictionary<string, string?>). We shouldn't be manipulating the options
that are given to us so the ReadOnly variants should be sufficient.

Use ImmutableDictionary<string, string?> internally for representing
attributes as it's easier to add and remove entries without modifying
the passed-in objects.